Use BITMAPINFOHEADER and write image to bmp
The `Bmp` class could only write an image filled with black of size n x
m. Now, it can write an image filled with arbitrary image data, as
specified by the red, green, and blue channels given by the user.

Also note that the use of BITMAPINFOHEADER is hard-coded, which may need
to change in the future.

The API in its current state is far from decent and will need a rework
in the near future, but what is here should work as a proof of concept.