Switch from redux to custom mutable state
As nice as redux is for traditional web apps, it recomputes many things
that will not be necessary for this web renderer, causing massive
performance drops. It also would not allow non-serializable objects in
the store, which means no `ArrayBuffer`.

With this update, one `ArrayBuffer` can be stored instead of many,
immutable js arrays. Tested on my system in development mode, and now a
4K image renders with absolutely no lag.